# bachelor/third year/Web_and_Software_Architecture/src/doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText API
  description: API for WASAText messaging platform
  version: 1.0.0
...
security: # Apply the security globally to all operations
  - bearerAuth: [] # use the same name as above
paths:
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/forwardMessage:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    post:
      tags:
        - message
      summary: Forward a message to a conversation
      description: Allows the user to forward an existing message to a conversation by specifying the conversation id in the request body.
      operationId: forwardMessage
      requestBody:
        description: Target conversation id for the forwarded message
        content:
          application/json:
            schema: 
              description: "Schema for forward message"
              type: object
              properties:
                conversationId:
                  description: Unique conversation identifier.
                  type: integer
                  example: 1
        required: true
      responses:
        "201":
          description: Message forward successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: Invalid conversation or message id, or invalid target conversation id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Original conversation or message not found, or target conversation not found
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/comments/{commentId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
      - $ref: "#/components/parameters/commentId"
    delete:
      tags:
        - message
      summary: Remove a comment from a message
      description: Allows the user to remove a specific comment from a message in a conversation.
      operationId: uncommentMessage
      responses:
        "204":
          description: Comment deleted successfully
        '400':
          description: Invalid message or comment id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404": 
          description: Conversation, message, or comment not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    delete:
      tags:
        - message
      summary: Delete a message from a conversation
      description: Allows the user to delete a specific message from a conversation.
      operationId: deleteMessage
      responses:
        "204":
          description: Message deleted successfully
        '400':
          description: Invalid message id or conversation id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404": 
          description: Conversation or message not found
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/groups/{groupId}/members/:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    put:
      tags:
        - group
      summary: Add a user to a group
      description: Allows the user to add another user to an existing group.
      operationId: addToGroup
      requestBody:
        description: User to add to the group
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        '204':
          description: User added successufully
        '400':
          description: Invalid group id or user id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group or user not found
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    get:
      tags:
        - group
      summary: Get group members
      description: Returns a list of usernames of users present in the specified group.
      operationId: getGroupMembers
      responses:
        '200':
          description: List of group members
          content:
            application/json:
              schema:
                description: "Schema for get group members"
                type: array
                minItems: 0
                maxItems: 10
                items:
                  $ref: "#/components/schemas/Username"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Group not found or user does not have access to the group
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/groups/{groupId}/members/me:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    delete:
      tags:
        - group
      summary: Leave a group
      description: Allows a user to leave an existing group they are part of.
      operationId: leaveGroup
      responses:
        "204":
          description: User successfully left the group
        '400':
          description: Invalid group id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        "404": 
          description: Group not found or user is not a member of the group
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /users/{userId}/search:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - user
      summary: Search for users by username
      description: Allows a user to search for other users by username.
      operationId: searchUsers
      parameters:
        - name: username
          in: query
          description: The username or part of the username to search for.
          required: true
          schema:
            $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: List of matching usernames
          content:
            application/json:
              schema:
                description: "Schema for set my photo response"
                type: array
                minItems: 0
                maxItems: 100
                items:
                  $ref: '#/components/schemas/Username'
        '400':
          description: Invalid search query
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes: # Define the security scheme type (HTTP bearer)
    bearerAuth: # arbitrary name for the security scheme
      type: http
      scheme: bearer
  schemas:
    Message:
      title: Message
      description: "Represent a single Message."
      type: object
      properties:
        id:
          description: Unique message identifier.
          type: integer
          example: 1
          readOnly: true
        timestamp: 
          type: string
          format: date-time
          description: Timestamp when the message was sent.
          example: "2023-11-05T14:30:00Z"
        content:
          description: Text or photo content.
          oneOf:
            - $ref: '#/components/schemas/Text'
            - $ref: '#/components/schemas/Photo'
        sender:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ["received", "read"]
          description: Status of the sent message indicated by checkmarks
          example: "read"
        commments:
          type: array
          minItems: 0
          maxItems: 500
          description: List of comment on the message
          items:
            $ref: '#/components/schemas/Comment'
        type:
          description: Type of the message
          type: string
          enum: ["standard", "forward", "reply"]
          example: "standard"
        replyToMessageId:
          type: integer
          description: The ID of the message this message is replying to.
          example: 12
    Comment:
      title: Comment
      description: "Represet a single Comment."
      type: object
      properties:
        commentId:
          description: Unique comment identifier.
          type: integer
          example: 1
          readOnly: true
        sender:
          $ref: '#/components/schemas/User'
        content:
          type: string
          description: The reaction as an emoji
          example: "👍"
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2
...
    User:
      title: User
      description: "Represents a single User."
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          $ref: '#/components/schemas/Username'
    Photo:
      description: Base64 encoded image data
      type: string
      format: byte
      pattern: "^[01]+$"
      minLength: 0
      maxLength: 9999999999999999999
    Username:
      description: Username of the user.
      type: string
      example: "user123"
      minLength: 3
      maxLength: 16
      pattern: "^.*?$"
    UserId:
      description: Unique user identifier
      type: integer
      example: 1
      readOnly: true
...
    Text:
      type: string
      description: Text content of the message
      example: "Hello, how are you?"
      minLength: 1
      maxLength: 500
      pattern: '^.*?$'
  parameters:
    userId:
      schema:
        $ref: '#/components/schemas/UserId'
      name: userId
      in: path
      required: true
      description: User id
    conversationId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier of the conversation."
      name: conversationId
      in: path
      required: true
      description: Conversation id
    messageId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the message."
      name: messageId
      in: path
      required: true
      description: Message id
    commentId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the comment."
      name: commentId
      in: path
      required: true
      description: Comment id
    groupId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Unique identifier for the group."
      name: groupId
      in: path
      required: true
      description: Group id
...
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
    InternalServerError:
      description: Internal server error